requirements.txt
flask==3.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
openai==0.28
eventlet==0.33.3
python-engineio==4.8.0
python-socketio==5.10.0